example/gifs_to_bin/pyproject.toml
[project]
name = "gifs-to-bin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]
